+++ src/awtk.c
@@ -31,6 +31,7 @@
 #include "base/input_method.h"
 #include "base/image_manager.h"
 #include "base/window_manager.h"
+#include "base/widget_animator_manager.h"
 #include "base/widget_factory.h"
 #include "base/assets_manager.h"
 #include "font_loader/font_loader_bitmap.h"
@@ -105,6 +106,8 @@ ret_t tk_init_internal(void) {
   return_value_if_fail(locale_info_set(locale_info_create(NULL, NULL)) == RET_OK, RET_FAIL);
   return_value_if_fail(font_manager_set(font_manager_create(font_loader)) == RET_OK, RET_FAIL);
   return_value_if_fail(image_manager_set(image_manager_create(image_loader)) == RET_OK, RET_FAIL);
+  return_value_if_fail(widget_animator_manager_set(widget_animator_manager_create()) == RET_OK,
+                       RET_FAIL);
   return_value_if_fail(window_manager_set(window_manager_create()) == RET_OK, RET_FAIL);
 
   return RET_OK;
@@ -120,6 +123,9 @@ ret_t tk_init(wh_t w, wh_t h, app_type_t app_type, const char* app_name, const c
 ret_t tk_deinit_internal(void) {
   widget_destroy(window_manager());
   window_manager_set(NULL);
+
+  widget_animator_manager_destroy(widget_animator_manager());
+  widget_animator_manager_set(NULL);
 
   idle_manager_destroy(idle_manager());
   idle_manager_set(NULL);