+++ src/base/widget_animator.h
@@ -28,17 +28,34 @@
 
 BEGIN_C_DECLS
 
+struct _widget_animator_manager_t;
+typedef struct _widget_animator_manager_t widget_animator_manager_t;
+
 struct _widget_animator_t;
 typedef struct _widget_animator_t widget_animator_t;
 
 typedef ret_t (*widget_animator_update_t)(widget_animator_t* animator, float_t percent);
 typedef ret_t (*widget_animator_destroy_t)(widget_animator_t* animator);
 
+typedef enum _animator_state_t {
+  ANIMATOR_CREATED = 0,
+  ANIMATOR_RUNNING,
+  ANIMATOR_PAUSED,
+  ANIMATOR_STOPPED,
+  ANIMATOR_DONE
+} animator_state_t;
+
 /**
  * @class widget_animator_t
  * 控件动画接口。
  */
 struct _widget_animator_t {
+  /**
+   * @property {char*} name
+   * 名称。
+   */
+  char* name;
+
   /**
    * @property {widget_t*} widget
    * @annotation ["private"]
@@ -75,6 +92,13 @@ struct _widget_animator_t {
   uint32_t now;
 
   /**
+   * @property {float_t} time_scale
+   * @annotation ["private"]
+   * 时间倍率。
+   */
+  float_t time_scale;
+
+  /**
    * @property {uint32_t} start_time
    * @annotation ["private"]
    * 起始时间(毫秒)。
@@ -96,6 +120,13 @@ struct _widget_animator_t {
   uint32_t duration;
 
   /**
+   * @property {animator_state_t} state
+   * @annotation ["private"]
+   * 动画的状态。
+   */
+  animator_state_t state;
+
+  /**
    * @property {uint32_t} emitter
    * @annotation ["private"]
    * emitter
@@ -115,13 +146,6 @@ struct _widget_animator_t {
    * 是否永远播放(yoyo_times/repeat_times为0时，自动设置此标志)。
    */
   bool_t forever;
-
-  /**
-   * @property {uint32_t} timer_id
-   * @annotation ["private"]
-   * 定时器ID。
-   */
-  uint32_t timer_id;
 
   /**
    * @property {uint32_t} widget_destroy_id
@@ -150,12 +174,16 @@ struct _widget_animator_t {
    * destroy函数，子类需要实现。
    */
   widget_animator_destroy_t destroy;
+
+  /*private*/
+  struct _widget_animator_t* next;
+  widget_animator_manager_t* widget_animator_manager;
 };
 
 /**
  * @method widget_animator_init
  * 初始化。仅供子类内部使用。
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @param {widget_animator_t*} animator 动画对象。
  * @param {widget_t*} widget 控件对象。
  * @param {uint32_t} duration 动画持续时间。
  * @param {uint32_t} delay 动画执行时间。
@@ -167,18 +195,19 @@ ret_t widget_animator_init(widget_animator_t* animator, widget_t* widget, uint32
                            uint32_t delay, easing_func_t easing);
 
 /**
- * @method widget_animator_start
- * 启动动画。
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @method widget_animator_set_name
+ * 设置名称。
+ * @param {widget_animator_t*} animator 动画对象。
+ * @param {const char*} name 名称。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
-ret_t widget_animator_start(widget_animator_t* animator);
+ret_t widget_animator_set_name(widget_animator_t* animator, const char* name);
 
 /**
  * @method widget_animator_set_yoyo
  * 设置为yoyo模式。
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @param {widget_animator_t*} animator 动画对象。
  * @param {uint32_t} yoyo_times yoyo的次数，往返视为两次。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
@@ -188,7 +217,7 @@ ret_t widget_animator_set_yoyo(widget_animator_t* animator, uint32_t yoyo_times)
 /**
  * @method widget_animator_set_repeat
  * 设置为重复模式。
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @param {widget_animator_t*} animator 动画对象。
  * @param {uint32_t} repeat_times 重复的次数。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
@@ -198,7 +227,7 @@ ret_t widget_animator_set_repeat(widget_animator_t* animator, uint32_t repeat_ti
 /**
  * @method widget_animator_set_reversed
  * 设置为逆向模式。
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @param {widget_animator_t*} animator 动画对象。
  * @param {bool_t} value 是否为逆向模式。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
@@ -206,12 +235,22 @@ ret_t widget_animator_set_repeat(widget_animator_t* animator, uint32_t repeat_ti
 ret_t widget_animator_set_reversed(widget_animator_t* animator, bool_t value);
 
 /**
+ * @method widget_animator_set_time_scale
+ * 设置时间倍率，用于实现时间加速减速和停滞的功能。
+ * @param {widget_animator_t*} animator 动画对象。
+ * @param {float_t} time_scale 时间倍率。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t widget_animator_set_time_scale(widget_animator_t* animator, float_t time_scale);
+
+/**
  * @method widget_animator_on
  * 注册指定事件的处理函数。
  * @annotation ["scriptable:custom"]
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @param {widget_animator_t*} animator 动画对象。
  * @param {event_type_t} type
- * 事件类型。目前支持：EVT_ANIM_START,EVT_ANIM_STOP,EVT_ANIM_ONCE和EVT_ANIM_END。
+ * 事件类型。目前支持：EVT_ANIM_START,EVT_ANIM_STOP,EVT_ANIM_PAUSE,EVT_ANIM_ONCE和EVT_ANIM_END。
  * @param {event_func_t} on_event 事件处理函数。
  * @param {void*} ctx 事件处理函数上下文。
  *
@@ -224,7 +263,7 @@ uint32_t widget_animator_on(widget_animator_t* animator, event_type_t type, even
  * @method widget_animator_off
  * 注销指定事件的处理函数。
  * @annotation ["scriptable:custom"]
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @param {widget_animator_t*} animator 动画对象。
  * @param {uint32_t} id widget_animator_on返回的ID。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
@@ -232,18 +271,36 @@ uint32_t widget_animator_on(widget_animator_t* animator, event_type_t type, even
 ret_t widget_animator_off(widget_animator_t* animator, uint32_t id);
 
 /**
+ * @method widget_animator_start
+ * 启动动画。
+ * @param {widget_animator_t*} animator 动画对象。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t widget_animator_start(widget_animator_t* animator);
+
+/**
  * @method widget_animator_stop
  * 停止动画。
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @param {widget_animator_t*} animator 动画对象。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
 ret_t widget_animator_stop(widget_animator_t* animator);
 
 /**
+ * @method widget_animator_pause
+ * 暂停动画。
+ * @param {widget_animator_t*} animator 动画对象。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t widget_animator_pause(widget_animator_t* animator);
+
+/**
  * @method widget_animator_set_destroy_when_done
- * 设置完成时是否自动销毁动画对象本身(缺省销毁)。
- * @param {widget_animator_t*} animator 动画对象本身。
+ * 设置完成时是否自动销毁动画对象(缺省销毁)。
+ * @param {widget_animator_t*} animator 动画对象。
  * @param {bool_t} destroy_when_done 完成时是否自动销毁动画对象。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
@@ -251,9 +308,20 @@ ret_t widget_animator_stop(widget_animator_t* animator);
 ret_t widget_animator_set_destroy_when_done(widget_animator_t* animator, bool_t destroy_when_done);
 
 /**
+ * @method widget_animator_time_elapse
+ * 时间流失，更新动画(由动画管理器调用)。
+ * @annotation ["private"]
+ * @param {widget_animator_t*} animator 动画对象。
+ * @param {uint32_t} delta_time 时间增量。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t widget_animator_time_elapse(widget_animator_t* animator, uint32_t delta_time);
+
+/**
  * @method widget_animator_destroy
  * 销毁animator对象。
- * @param {widget_animator_t*} animator 动画对象本身。
+ * @param {widget_animator_t*} animator 动画对象。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */