+++ tests/timer_test.cc
@@ -24,14 +24,6 @@ static ret_t timer_once(const timer_info_t* timer) {
   return RET_OK;
 }
 
-static ret_t timer_delta_time(const timer_info_t* timer) {
-  char str[32];
-  tk_snprintf(str, sizeof(str), "%d:%d", timer->now, timer->delta_time);
-  s_log += str;
-
-  return RET_REPEAT;
-}
-
 static ret_t timer_repeat(const timer_info_t* timer) {
   s_log += "r:";
   return RET_REPEAT;
@@ -90,26 +82,6 @@ TEST(Timer, once) {
   ASSERT_EQ(timer_manager_dispatch(tm), RET_OK);
   ASSERT_EQ(timer_manager_count(tm), 0);
   ASSERT_EQ(s_log, repeat_str("o:", NR));
-
-  timer_manager_destroy(tm);
-}
-
-TEST(Timer, delta_time) {
-  timer_manager_t* tm = NULL;
-
-  timer_set_time(80);
-  tm = timer_manager_create(timer_get_time);
-
-  timer_clear_log();
-  timer_manager_add(tm, timer_delta_time, NULL, 10);
-  timer_set_time(100);
-  ASSERT_EQ(timer_manager_dispatch(tm), RET_OK);
-  ASSERT_EQ(s_log, "100:20");
-
-  timer_clear_log();
-  timer_set_time(118);
-  ASSERT_EQ(timer_manager_dispatch(tm), RET_OK);
-  ASSERT_EQ(s_log, "118:18");
 
   timer_manager_destroy(tm);
 }